refactor: restructured checkbox sections into a single reusable component
