--- src/components/WrapperComponent/WrapperComponent.module.scss
@@ -1,11 +1,0 @@
-.wrapper {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  align-items: center;
-  gap: 50px;
-}
-
-.carousel {
-  margin-top: 3.5px;
-}
